ajuste na pagina de detalhe e inclusão do databinding

+++ app/src/main/java/br/com/homartapp/data/remote/LocationResponse.kt
@@ -8,5 +8,5 @@ data class Location(
     val name: String,
     val review: Double,
     val type: String,
-    val id: Long
+    val id: Int
 )
